Fix BattleDetailPage,  article instead of main, and rename folder name
