+++ src/battle/pages/BattleDetailPage/BattleDetailPage.css
