+++ src/battle/pages/BattleDetailPage/BattleDetailPage.test.tsx
