+++ src/router/TestAppRouter.tsx
@@ -1,7 +1,7 @@
 import { Route, Routes } from "react-router";
 import NotFoundPage from "../pages/NotFoundPage.tsx/NotFoundPage";
 import BattlesPage from "../battle/pages/BattlesPage/BattlesPage";
-import BattleDetailPage from "../battle/pages/BattleDetail/BattleDetailPage";
+import BattleDetailPage from "../battle/pages/BattleDetailPage/BattleDetailPage";
 
 const AppRouterTest: React.FC = () => {
   return (
